added new inner function variant
added a new variant of the curve stacking function. this takes about 0.4 as long as the previous version, though i think the other implementation (curve_stack_3x3_xy) is worth preserving. it's a clearer reflection of how the algorithm works, and might be desirable in cases where fewer samples with better coverage are preferable.